requirements.txt
streamlit
pandas
google-cloud-firestore
altair